Added Java-created tracings for Sections 46 and 47.
These are the first tracings generated from the Java version of
Reconstruct. They suffer from a number of problems:

  - Object names come from the currently selected color:
        RGB_100 (red=1, green=0, blue=0)
        RGB_010 (red=0, green=1, blue=0)

  - Traces can only be made with polygon segments (no curves)

There may be other problems with these traces, but they were able to be
imported into Reconstruct and appeared to be displayed properly.

If these problems can't be easily fixed (by editing the XML, for
example), please feel free to delete them.