Fixed XML to use consistent naming and colors for new Java traces.
The current Java version uses the trace color to make up a name. This
results in names like RGB_100 and RGB_010. These were updated to the
current temporary name "bob001" to be consistent with the earlier traces
made in Reconstruct.exe. The colors were also made consistent (0,0.5,0)
for the "bob001" object.

The use of personal names is being tested as a way to avoid conflicts in
collaboratively traced series. It avoids two people happening to choose
"d23" for the same object (assuming everyone has a different name/id).
User names (in GitHub, for example) might be used for this purpose.

Alternatively, names could be brokered through the README.md file where
names are claimed or declared as part of the check out/in process.