Finished tracing d01 through all sections with Java version.
As a test of the code, I finished tracing d01 through the remaining
sections. However, this version contains a mistake that the Java tool
can't fix (it doesn't have editing abilities yet). So this version will
be cleaned up with the Windows version of Reconstruct. The new version
will be added as a new commit.