Added an animation showing updates from previous commit.
Note that the image file is named Git_Commit_### to show the series at
that state. This may not be such a useful convention since using a
generic name (such as "series.gif") would contain the history of all the
commits and could be viewed at any stage of the process. This would be a
good decision for the "Series Captain" to make. :)

Note also that this commit was traced using the current version of the
Java (Git commit d05a04ab25273180ad9...) on a Linux system.