Added a Bezier trace to section 1 (shows up as section 2).
The current version of Reconstruct Java can generate Bezier traces, but
it's a little bit clunky. The mode is enabled through the Extras/Mode
menu by checking both "Segment Drawing" and "Bezier Drawing". Traces are
then drawn click by click, but the segments will show up as straight
lines (not curves). But when the file is saved and then opened again,
the trace will be drawn as a Bezier curve with handles (which can't be
used for editing yet).

This commit shows a Bezier trace just above the mid point of the first
non-calibration section. It's a red trace just below and to the left of
the red trace marked with an "X". It's also above the yellow trace
marked with an octogon.

Note that the "handles" for these traces are stored in the XML in a
field that Reconstruct.exe doesn't know about or use. The trace will be
retained after opening with Reconstruct.exe as long as that section
isn't rewritten by Reconstruct.exe. If Reconstruct.exe re-rewites that
section, then the "handles" will be lost and the trace will again become
a normal polygon trace.